Update README.md
Codoffee Quiz Maker is the first project that we did years ago in our startup codoffee(which dismissed in early 2020) I did parts of it as i was 13-y.o by the time it was being done for our schools celebration(competetion actually) and it won. In this repository i'll present you with it's full codec + how to run it + the full manual of software.